requirements.txt
soundfile
pyttsx3
google-generativeai
speechrecognition
selenium
pyaudio
python-dotenv
